remove extraneous tabs, and extra crazy tests
