+++ shopping-list/app/(tabs)/index.delete.test.tsx
@@ -10,7 +10,7 @@
 import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react-native';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import ShoppingListsScreen from './shopping-lists';
+import ShoppingListsScreen from './index';
 import { shoppingListsApi } from '@/api/client';
 
 // Mock the API client
